Move forward declaration to own header file
